+++ packages/article-skeleton/src/article-body/article-body.js
@@ -65,6 +65,9 @@ import {
 
 const deckApiUrl = "https://gobble.timesdev.tools/deck/api/deck-post-action/";
 
+const disabledAds = ["c8bf6998-d498-11ed-b5c3-54651fc826e9"];
+const hasDisabledAds = id => disabledAds.includes(id);
+
 export const responsiveDisplayWrapper = displayType => {
   switch (displayType) {
     case "secondary":
@@ -100,11 +103,12 @@ const renderers = ({
   isPreview,
   isLiveOrBreaking,
   section,
-  articleHeadline
+  articleHeadline,
+  articleId
 }) => ({
   ...coreRenderers,
   ad(key) {
-    return (
+    return hasDisabledAds(articleId) ? null : (
       <InlineAdWrapper>
         <InlineAdTitle>Advertisement</InlineAdTitle>
         <AdContainer key={key} slotName="inline-ad" />
@@ -112,7 +116,7 @@ const renderers = ({
     );
   },
   inlineAd1(key) {
-    return (
+    return hasDisabledAds(articleId) ? null : (
       <InlineAdWrapper>
         <InlineAdTitle>Advertisement</InlineAdTitle>
         <AdContainer key={key} slotName="inlineAd1" />
@@ -120,7 +124,7 @@ const renderers = ({
     );
   },
   inlineAd2(key) {
-    return (
+    return hasDisabledAds(articleId) ? null : (
       <InlineAdWrapper>
         <InlineAdTitle>Advertisement</InlineAdTitle>
         <AdContainer key={key} slotName="inlineAd2" />
@@ -128,7 +132,7 @@ const renderers = ({
     );
   },
   inlineAd3(key) {
-    return (
+    return hasDisabledAds(articleId) ? null : (
       <InlineAdWrapper>
         <InlineAdTitle>Advertisement</InlineAdTitle>
         <AdContainer key={key} slotName="inlineAd3" />
@@ -149,7 +153,7 @@ const renderers = ({
     );
   },
   nativeAd(key) {
-    return isLiveOrBreaking ? null : (
+    return isLiveOrBreaking || hasDisabledAds(articleId) ? null : (
       <NativeAd className="group-3 hidden" key={key}>
         <NativeAdTitle>Sponsored</NativeAdTitle>
         <Ad id="advert-inarticle-native-1" data-parent="group-3" />
@@ -575,7 +579,8 @@ const ArticleBody = ({
   swgProductId,
   inArticlePuffFlag,
   isLiveOrBreaking,
-  articleHeadline
+  articleHeadline,
+  id: articleId
 }) =>
   renderTrees(
     bodyContent.map(decorateAd({ contextUrl, section })),
@@ -586,6 +591,7 @@ const ArticleBody = ({
       swgProductId,
       inArticlePuffFlag,
       isLiveOrBreaking,
+      articleId,
       section,
       articleHeadline
     })