+++ packages/ts-newskit/src/components/slices/shared-styles/index.ts
@@ -1,5 +1,10 @@
-import { styled, TextBlock } from 'newskit';
+import { styled, TextBlock, Block, getSizingCssFromTheme } from 'newskit';
 
 export const ColouredText = styled(TextBlock)<{ $color?: string }>`
   ${({ $color }) => $color && `color: ${$color};`};
+`;
+
+export const ContainerInline = styled(Block)`
+  display: inline-block;
+  ${getSizingCssFromTheme('height', 'sizing020')};
 `;